requirements.txt
numpy==1.23.2
Pillow==9.0.1
